backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
websockets==12.0
aiohttp==3.9.1
web3==6.12.0
pyyaml==6.0.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
prometheus-client==0.19.0
python-telegram-bot==20.7
slack-sdk==3.26.1 